Window moving now works 100%
Added window moving support, complete with the ability to click-and-
drag windows. The contents do not update during move as of yet, but
the previous limitation requiring the mouse to remain within the
window has been removed.

This version logs a crap load to the console, and isn't final. Once
the window model is complete, it should become less verbose.