Added test window to start menu, hidden by default
Hid test window by default, preferring to add it to the start menu
from where it can be invoked at any time. This means that it can
be created multiple times in a row.

Note that this breaks closing right now, so only one should be
created at a time.