added some much-needed style elements
Added dropshadow so windows on top of windows can be differentiated

Also added styling to the close button so it doesn't jutt out so
badly with KDE's breeze theme and look so out of place otherwise.

Also added addition to the blank window to make the menu close when
it is opened.