Added grabhandles for resizing
Resizing a window is a feature that is not so far off in the future
so we should be prepared by adding a grab handle so that it can be
done by the user later. Thie images have been generated and tested
to ensure the browser does not drag and drop the image instead of
resizing the window.