Switched all toplevel element to window object
Changed all signals connected to just the toplevel to connect to
the whole window object. This fixes a bug where window movement
would be broken because the window object was required at the same
time as the window toplevel dom node in the same variable.

So yes, I created this bug in the last commit, but switching
everything to using the window object instead of a dom node makes
sense because in the future, if something needs to be accessed
somewhere odd, this can be accomplished.