Added windowbody overflow property
The window body would allow text to overflow a resized window,
making it look like the window was not containing the text, merely
serving as a backdrop for it. This new style clips the text off the
side when there isn't enough space. In the future, it may be worth
looking into setting this to overflow:auto instead in order to
accomodate situations where there is content which won't fit on a
smaller screen.