Fixed long-length panel buttons
Windows with a long title length would push other window buttons
down to the last line of their text (bottom) in browsers like
Firefox. In browsers like Chrome this wouldn't be an issue, but
making this work in Firefox is of course important becuase it's the
browser this whole site has been developed on so far.