More composer view example updates
Composer view continues to be updated with more example content.
It has become clear at this point that it is a better idea to use
a <div contenteditable> tag instead of a plain textedit tag in
order to make the composer view, because this would allow for a
little more formatting freedom.