Added widgetTools.js to act as widget toolkit
Now that a windowing toolkit is complete, there needs to be one
widget toolkit which can be used anywhere to script elements
dynamically.

This is not such an issue in a case where there can only be one
copy of a widget, such as a static web interface. In places where
things need to be dynamically created and destroyed, and there is
no way of knowing how many instances of an object will exist, a
toolkit as such becomes a requirement.