Started rewriting the composer using widgetTools
Widgettools is now more or less ready to start driving basic buttons
and windows in a simple manner. It is not ready for anything that
actually does anything, as it still has no signal connector, but it
has enough that it can create and track its own widgets, including
state changes.