Continued porting to WidgetTools
WindowTools now has a place to add signal handlers for resize,
allowing widget toolkits to make use of that for full-size widgets
that don't do any CSS assery where they fill more space than they
should due to a mix of absolute and relative pixel space

The composer now has all the elements it had last night, they're
just created through the toolkit instead of on their own. This
means that there is the potential to toggle every single button on
its own, instead of making only the first to be created react, like
last night. This is because the toggled state won't call an object
by id, it will do so straight through JS DOM.