Added login screen mockup, composer login added
A mockup for a simple 640x480 login screen has been added. The aim is to
eventually make this kind of design be the standard for how registered
blog users sign in, showing up in the composer view window. If they
manage to successfully authenticate, they will be allowed into the main
composer view previously featured here.

The window itself is under construction, and for whatever reason will
not launch right now, but that is down the list and will be fixed soon.