Added toolbar to test window, added HR to toolkit
widgetTools now features a way to create horizontal rules in a window.

The test window now features not only the HR, but also a toolbar. These
have been used elsewhere, but it is important to feature them in this
test environment to make sure they don't break. The top button has been
converted to a tbutton, and a second one has been added alongside it.