Restructured directories
Directory structure before (which was better than nothing) done by
kroche was OK, but it didn't compartmentalize components with their own
resources, it split the whole project as if each file was built to be
dependant on the project. This isn't wrong, but it didn't fit with how
widgetTools and windowTools were designed. Each one was supposed to be a
drag-and-drop module which can be put anywhere.