Moved all JS sources to the index.js file
There was a sizeable amount of JavaScript inside the HTML index file.
This is not directly an issue, but for organization's sake this has been
moved to the similarly named index.js in the resources folder. Since it
was originally designed to contain scripts specific to index.html, there
is little or no possibility for anything to break.