Added login box, removed old menu items
The composer window now opens up to a nice login box, where users enter
their username and password. Right now the majority of that does not do
anything, simply acting as a placeholder for when the actual
functionality is ready to go. It does not check the values of the fields
for user and password, it simply opens the composer window when you hit
login.

The menu had a couple very old menu buttons, one for a blank window, and
one for a test window. Since these were mainly needed in order to make
the old window toolkit tests go more smoothly and have no value, real or
otherwise, they have been removed. Any of the actually functioning
windows now available can serve as testing for windowTools, including
the test window for WidgetTools.