Commented libcookie
libcookie now contains the same standard comment format as the other
portions of the interface. It is designed to act as API documentation.

One improvement which will eventually be applied to the rest of the
comments:
Data types are specified with arguments and returns in order to avoid
unexpected exceptions in the code.