Fixed random number checker
The random number checker did not accept any random string, in the
database or otherwise. It now only accepts ones that are in the database
from within the certain time frame allowed in the code.

Next step is to make it update the "last hit" timestamp so users can
continue working without interruptions