Renamed composerview.js to backendinterface.js
This is to reflect the fact that this file is built to also do authentication
work as well as composition of posts. There are a few other little JS files
that should eventually be migrated into larger categorical files. This will
allow for better structure and organization for the upcoming 0.1 release.