Created wallpaper folder. This commit breaks wallpaper.
The wall folder in resources was created. Wallpapers were then moved to
that folder.

This commit breaks wallpapers on the system as it is right now, since
the mechanism for switching wallpapers has not been updated as it is.

An upcoming commit will bring the wallpaper system inline with the new
changes, as well as streamline the selection backend.