Moved cookie stuff to libCookie
libCookie had an inferior grabCookie compared to index.js's getCookie,
unfortunately, neither is really the best description, so the function
was changed to readCookie, placed in libcookie.js with the good code,
and all references in index.js were changed to match the new function
name.

The unused setCookie() function in index.js was also removed.