Updated centerWindow function
Fixed some pixel maths that caused the window to center to a position
offscreen. It now (approximately) centers to the actual center of the
screen.

Because centerWindow() exists and works properly, the window positioning
argument of addWindow() may soon be removed, as it makes no sense to
center an empty window only to add contents.